--- frontend/src/colorModeToggler.js
@@ -1,89 +1,0 @@
-(() => {
-  const getStoredTheme = () => localStorage.getItem("theme");
-  const setStoredTheme = (theme) => localStorage.setItem("theme", theme);
-
-  const getPreferredTheme = () => {
-    const storedTheme = getStoredTheme();
-    if (storedTheme) {
-      return storedTheme;
-    }
-
-    return window.matchMedia("(prefers-color-scheme: dark)").matches
-      ? "dark"
-      : "light";
-  };
-
-  const setTheme = (theme) => {
-    if (
-      theme === "auto" &&
-      window.matchMedia("(prefers-color-scheme: dark)").matches
-    ) {
-      document.documentElement.setAttribute("data-bs-theme", "dark");
-    } else {
-      document.documentElement.setAttribute("data-bs-theme", theme);
-    }
-  };
-
-  setTheme(getPreferredTheme());
-
-  const showActiveTheme = (theme, focus = false) => {
-    const themeSwitcher = document.querySelector("#bd-theme");
-
-    console.log("in showActiveTheme");
-    console.log(themeSwitcher);
-
-    if (!themeSwitcher) {
-      return;
-    }
-
-    const themeSwitcherText = document.querySelector("#bd-theme-text");
-    const activeThemeIcon = document.querySelector("i.theme-icon-active");
-    const btnToActive = document.querySelector(
-      `[data-bs-theme-value="${theme}"]`
-    );
-    const iconOfActiveBtn = btnToActive
-      .querySelector("i")
-      .getAttribute("data-icon");
-
-    document.querySelectorAll("[data-bs-theme-value]").forEach((element) => {
-      element.classList.remove("active");
-      element.setAttribute("aria-pressed", "false");
-    });
-
-    btnToActive.classList.add("active");
-    btnToActive.setAttribute("aria-pressed", "true");
-    let currentIcon = activeThemeIcon.getAttribute("data-active-icon");
-    activeThemeIcon.classList.remove(`bi-${currentIcon}`);
-    activeThemeIcon.setAttribute("data-active-icon", iconOfActiveBtn);
-    activeThemeIcon.classList.add(`bi-${iconOfActiveBtn}`);
-    const themeSwitcherLabel = `${themeSwitcherText.textContent} (${btnToActive.dataset.bsThemeValue})`;
-    themeSwitcher.setAttribute("aria-label", themeSwitcherLabel);
-
-    if (focus) {
-      themeSwitcher.focus();
-    }
-  };
-
-  window
-    .matchMedia("(prefers-color-scheme: dark)")
-    .addEventListener("change", () => {
-      const storedTheme = getStoredTheme();
-      if (storedTheme !== "light" && storedTheme !== "dark") {
-        setTheme(getPreferredTheme());
-      }
-    });
-
-  window.addEventListener("DOMContentLoaded", () => {
-    console.log("in DOMContentLoaded");
-    showActiveTheme(getPreferredTheme());
-
-    document.querySelectorAll("[data-bs-theme-value]").forEach((toggle) => {
-      toggle.addEventListener("click", () => {
-        const theme = toggle.getAttribute("data-bs-theme-value");
-        setStoredTheme(theme);
-        setTheme(theme);
-        showActiveTheme(theme, true);
-      });
-    });
-  });
-})();
