+++ frontend/src/components/Output.jsx
