+++ frontend/src/components/Prompt.jsx
