+++ frontend/src/components/ResizeableTile.jsx
