+++ frontend/src/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 
 // Bootstrap CSS
@@ -11,8 +10,9 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap/dist/js/bootstrap.bundle.min";
 //
 import "react-grid-layout/css/styles.css";
-import "./App.css";
-import "./custom.scss";
+import "./css/App.css";
+import "./css/index.css";
+import "./css/scss/custom.scss";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
